Revert "Add gp ebs support to `aws_ec2` + Let ec2 tests run multi region (#916)"
This reverts commit 0325bb9f4b2af03c7f301ef1ef5b9b7447365c15.
