+++ src/modules/0.0.11/aws_ec2/entity/index.ts
@@ -1,3 +1,2 @@
-export * from './general_purpose_volume'
 export * from './instance'
 export * from './registered_instance'
